Add a `definitions/.cli-metadata.json` file for use by the CLI
This file will be retrieved and consulted by the CLI before doing any relevant
interactions with the GitHub repo. It's purpose is to contain any necessary
"latest" information needed for the CLI to do use repository data.

At the moment, this only includes a "minimumCLIVersion" field -- which tells the
CLI to bail out (and inform the user) if its own version is below a threshold.
This is mostly a fail-safe in case we decide to change the structure of the repo
at any point in a way that isn't backward compatible with existing clients.