Add shelljs declarations (#189)
This PR adds flow declarations for `shelljs`. It covers the complete API but in some cases I have restricted the interfaces for certain functions to the part that can be reasonably typed. However, this should not limit functionality.

I've added tests following some of the others as examples but I had some trouble getting the `flow-typed` CLI tool to run them.