Add definition for aphrodite.
This adds a definition for the aphrodite styles-in-js library. It
functions all the way back to flow 28 and I targeted v0.5 in my
definition. A couple places annotate simply `Object` as I ran into some
issues with more specific types conforming to the expected shape. Let me
know if something more specific is prefered and I can try my best to
make that happen :)