Fix prettier errors throw during `./run_def_tests.sh` (#1205)
* Fix prettier errors throw during `./run_def_tests.sh`

A number of prettier errors were thrown when running
`./run_def_tests.sh`. This fixes those.

Also when running said command the `definitions/package-lock.json` was
updated with some new versions. I'm not sure why that happened. I can
remove that change if needed.

* Redo commit but without `precommit` script

The `precommit` script is modifying files with the wrong version of prettier which is breaking code changes.

* Revert "Redo commit but without `precommit` script"

This reverts commit 87d2724de17a4ac14f7a8c704173720b49fa0be3.

* Revert "Fix prettier errors throw during `./run_def_tests.sh`"

This reverts commit 5bd195bc0dafde5af5b9f275e439ddd4c216fed1.

* Use `yarn` in `run_def_tests` for `cli/` instead of `npm`

Doing `npm install` inside of `run_def_tests` is different to how CI runs tests which was leading to tests failing locally that would pass on CI.