Reenable Ramda's `sortBy` (#1243)
`sortBy` was disabled due to failing tests after v0.38. This was because
flow started to correctly infer a type error in the test.

The test was using the `nth` function which expects an Array but the
output of the `sortBy` test was expecting a Tuple. Flow does not
consider Tuples and Array equal and threw a type check error even though
this is completely valid at runtime.

Since Flow favors soundness, this is a valid type error. The fix is
re-enabling `sortBy` and fixing the test to not use `nth`.